--- blocks/code-highlighted/dark.css
@@ -1,177 +1,0 @@
-/*!
-  Theme: Default Dark
-  Author: Chris Kempson (http://chriskempson.com)
-  License: ~ MIT (or more permissive) [via base16-schemes-source]
-  Maintainer: @highlightjs/core-team
-  Version: 2021.09.0
-*/
-
-/*
-  WARNING: DO NOT EDIT THIS FILE DIRECTLY.
-
-  This theme file was auto-generated from the Base16 scheme default-dark
-  by the Highlight.js Base16 template builder.
-
-  - https://github.com/highlightjs/base16-highlightjs
-*/
-
-/*
-base00  #181818  Default Background
-base01  #282828  Lighter Background (Used for status bars, line number and folding marks)
-base02  #383838  Selection Background
-base03  #585858  Comments, Invisibles, Line Highlighting
-base04  #b8b8b8  Dark Foreground (Used for status bars)
-base05  #d8d8d8  Default Foreground, Caret, Delimiters, Operators
-base06  #e8e8e8  Light Foreground (Not often used)
-base07  #f8f8f8  Light Background (Not often used)
-base08  #ab4642  Variables, XML Tags, Markup Link Text, Markup Lists, Diff Deleted
-base09  #dc9656  Integers, Boolean, Constants, XML Attributes, Markup Link Url
-base0A  #f7ca88  Classes, Markup Bold, Search Text Background
-base0B  #a1b56c  Strings, Inherited Class, Markup Code, Diff Inserted
-base0C  #86c1b9  Support, Regular Expressions, Escape Characters, Markup Quotes
-base0D  #7cafc2  Functions, Methods, Attribute IDs, Headings
-base0E  #ba8baf  Keywords, Storage, Selector, Markup Italic, Diff Changed
-base0F  #a16946  Deprecated, Opening/Closing Embedded Language Tags, e.g. <?php ?>
-*/
-
-.code-highlighted.dark pre code.hljs {
-  display: block;
-  overflow-x: auto;
-  padding: 1em;
-}
-
-.code-highlighted.dark code.hljs {
-  padding: 3px 5px;
-}
-
-.code-highlighted.dark .hljs {
-  color: #d8d8d8;
-  background: #181818;
-}
-
-.code-highlighted.dark .hljs::selection,
-.code-highlighted.dark .hljs ::selection {
-  background-color: #383838;
-  color: #d8d8d8;
-}
-
-
-/* purposely do not highlight these things */
-.code-highlighted.dark .hljs-formula,
-.code-highlighted.dark .hljs-params,
-.code-highlighted.dark .hljs-property
-{}
-
-/* base03 - #585858 -  Comments, Invisibles, Line Highlighting */
-.code-highlighted.dark .hljs-comment {
-  color: #585858;
-}
-
-/* base04 - #b8b8b8 -  Dark Foreground (Used for status bars) */
-.code-highlighted.dark .hljs-tag {
-  color: #b8b8b8;
-}
-
-/* base05 - #d8d8d8 -  Default Foreground, Caret, Delimiters, Operators */
-.code-highlighted.dark .hljs-subst,
-.code-highlighted.dark .hljs-punctuation,
-.code-highlighted.dark .hljs-operator {
-  color: #d8d8d8;
-}
-
-.code-highlighted.dark .hljs-operator {
-  opacity: 0.7;
-}
-
-/* base08 - Variables, XML Tags, Markup Link Text, Markup Lists, Diff Deleted */
-.code-highlighted.dark .hljs-bullet,
-.code-highlighted.dark .hljs-variable,
-.code-highlighted.dark .hljs-template-variable,
-.code-highlighted.dark .hljs-selector-tag,
-.code-highlighted.dark .hljs-name,
-.code-highlighted.dark .hljs-deletion {
-  color: #ab4642;
-}
-
-/* base09 - Integers, Boolean, Constants, XML Attributes, Markup Link Url */
-.code-highlighted.dark .hljs-symbol,
-.code-highlighted.dark .hljs-number,
-.code-highlighted.dark .hljs-link,
-.code-highlighted.dark .hljs-attr,
-.code-highlighted.dark .hljs-variable.constant_,
-.code-highlighted.dark .hljs-literal {
-  color: #dc9656;
-}
-
-/* base0A - Classes, Markup Bold, Search Text Background */
-.code-highlighted.dark .hljs-title,
-.code-highlighted.dark .hljs-class .hljs-title,
-.code-highlighted.dark .hljs-title.class_
-{
-  color: #f7ca88;
-}
-
-.code-highlighted.dark .hljs-strong {
-  font-weight:bold;
-  color: #f7ca88;
-}
-
-/* base0B - Strings, Inherited Class, Markup Code, Diff Inserted */
-.code-highlighted.dark .hljs-code,
-.code-highlighted.dark .hljs-addition,
-.code-highlighted.dark .hljs-title.class_.inherited__,
-.code-highlighted.dark .hljs-string {
-  color: #a1b56c;
-}
-
-/* base0C - Support, Regular Expressions, Escape Characters, Markup Quotes */
-.code-highlighted.dark .hljs-built_in,
-.code-highlighted.dark .hljs-doctag, /* guessing */
-.code-highlighted.dark .hljs-quote,
-.code-highlighted.dark .hljs-keyword.hljs-atrule,
-.code-highlighted.dark .hljs-regexp {
-  color: #86c1b9;
-}
-
-/* base0D - Functions, Methods, Attribute IDs, Headings */
-.code-highlighted.dark .hljs-function .hljs-title,
-.code-highlighted.dark .hljs-attribute,
-.code-highlighted.dark .ruby .hljs-property,
-.code-highlighted.dark .hljs-title.function_,
-.code-highlighted.dark .hljs-section {
-  color: #7cafc2;
-}
-
-/* base0E - Keywords, Storage, Selector, Markup Italic, Diff Changed */
-.code-highlighted.dark .hljs-type,
-/* .hljs-selector-id, */
-/* .hljs-selector-class, */
-/* .hljs-selector-attr, */
-/* .hljs-selector-pseudo, */
-.code-highlighted.dark .hljs-template-tag,
-.code-highlighted.dark .diff .hljs-meta,
-.code-highlighted.dark .hljs-keyword {
-  color: #ba8baf;
-}
-.code-highlighted.dark .hljs-emphasis {
-  color: #ba8baf;
-  font-style: italic;
-}
-
-/* base0F - Deprecated, Opening/Closing Embedded Language Tags, e.g. <?php ?> */
-.code-highlighted.dark .hljs-meta,
-/*
-  prevent top level .keyword and .string scopes
-  from leaking into meta by accident
-*/
-.code-highlighted.dark .hljs-meta .hljs-keyword,
-.code-highlighted.dark .hljs-meta .hljs-string
-{
-  color: #a16946;
-}
-
-.code-highlighted.dark .hljs-meta .hljs-keyword,
-/* for v10 compatible themes */
-.code-highlighted.dark .hljs-meta-keyword {
-  font-weight: bold;
-}
