--- blocks/code-highlighted/default.css
@@ -1,177 +1,0 @@
-/*!
-  Theme: Default Light
-  Author: Chris Kempson (http://chriskempson.com)
-  License: ~ MIT (or more permissive) [via base16-schemes-source]
-  Maintainer: @highlightjs/core-team
-  Version: 2021.09.0
-*/
-
-/*
-  WARNING: DO NOT EDIT THIS FILE DIRECTLY.
-
-  This theme file was auto-generated from the Base16 scheme default-light
-  by the Highlight.js Base16 template builder.
-
-  - https://github.com/highlightjs/base16-highlightjs
-*/
-
-/*
-base00  #f8f8f8  Default Background
-base01  #e8e8e8  Lighter Background (Used for status bars, line number and folding marks)
-base02  #d8d8d8  Selection Background
-base03  #b8b8b8  Comments, Invisibles, Line Highlighting
-base04  #585858  Dark Foreground (Used for status bars)
-base05  #383838  Default Foreground, Caret, Delimiters, Operators
-base06  #282828  Light Foreground (Not often used)
-base07  #181818  Light Background (Not often used)
-base08  #ab4642  Variables, XML Tags, Markup Link Text, Markup Lists, Diff Deleted
-base09  #dc9656  Integers, Boolean, Constants, XML Attributes, Markup Link Url
-base0A  #f7ca88  Classes, Markup Bold, Search Text Background
-base0B  #a1b56c  Strings, Inherited Class, Markup Code, Diff Inserted
-base0C  #86c1b9  Support, Regular Expressions, Escape Characters, Markup Quotes
-base0D  #7cafc2  Functions, Methods, Attribute IDs, Headings
-base0E  #ba8baf  Keywords, Storage, Selector, Markup Italic, Diff Changed
-base0F  #a16946  Deprecated, Opening/Closing Embedded Language Tags, e.g. <?php ?>
-*/
-
-.code-highlighted:not(.dark) pre code.hljs {
-  display: block;
-  overflow-x: auto;
-  padding: 1em;
-}
-
-.code-highlighted:not(.dark) code.hljs {
-  padding: 3px 5px;
-}
-
-.code-highlighted:not(.dark) .hljs {
-  color: #383838;
-  background: var(--overlay-background-color);
-}
-
-.code-highlighted:not(.dark) .hljs::selection,
-.code-highlighted:not(.dark) .hljs ::selection {
-  background-color: #d8d8d8;
-  color: #383838;
-}
-
-
-/* purposely do not highlight these things */
-.code-highlighted:not(.dark) .hljs-formula,
-.code-highlighted:not(.dark) .hljs-params,
-.code-highlighted:not(.dark) .hljs-property
-{}
-
-/* base03 - #b8b8b8 -  Comments, Invisibles, Line Highlighting */
-.code-highlighted:not(.dark) .hljs-comment {
-  color: #b8b8b8;
-}
-
-/* base04 - #585858 -  Dark Foreground (Used for status bars) */
-.code-highlighted:not(.dark) .hljs-tag {
-  color: #585858;
-}
-
-/* base05 - #383838 -  Default Foreground, Caret, Delimiters, Operators */
-.code-highlighted:not(.dark) .hljs-subst,
-.code-highlighted:not(.dark) .hljs-punctuation,
-.code-highlighted:not(.dark) .hljs-operator {
-  color: #383838;
-}
-
-.code-highlighted:not(.dark) .hljs-operator {
-  opacity: 0.7;
-}
-
-/* base08 - Variables, XML Tags, Markup Link Text, Markup Lists, Diff Deleted */
-.code-highlighted:not(.dark) .hljs-bullet,
-.code-highlighted:not(.dark) .hljs-variable,
-.code-highlighted:not(.dark) .hljs-template-variable,
-.code-highlighted:not(.dark) .hljs-selector-tag,
-.code-highlighted:not(.dark) .hljs-name,
-.code-highlighted:not(.dark) .hljs-deletion {
-  color: #ab4642;
-}
-
-/* base09 - Integers, Boolean, Constants, XML Attributes, Markup Link Url */
-.code-highlighted:not(.dark) .hljs-symbol,
-.code-highlighted:not(.dark) .hljs-number,
-.code-highlighted:not(.dark) .hljs-link,
-.code-highlighted:not(.dark) .hljs-attr,
-.code-highlighted:not(.dark) .hljs-variable.constant_,
-.code-highlighted:not(.dark) .hljs-literal {
-  color: #dc9656;
-}
-
-/* base0A - Classes, Markup Bold, Search Text Background */
-.code-highlighted:not(.dark) .hljs-title,
-.code-highlighted:not(.dark) .hljs-class .hljs-title,
-.code-highlighted:not(.dark) .hljs-title.class_
-{
-  color: #f7ca88;
-}
-
-.code-highlighted:not(.dark) .hljs-strong {
-  font-weight:bold;
-  color: #f7ca88;
-}
-
-/* base0B - Strings, Inherited Class, Markup Code, Diff Inserted */
-.code-highlighted:not(.dark) .hljs-code,
-.code-highlighted:not(.dark) .hljs-addition,
-.code-highlighted:not(.dark) .hljs-title.class_.inherited__,
-.code-highlighted:not(.dark) .hljs-string {
-  color: #a1b56c;
-}
-
-/* base0C - Support, Regular Expressions, Escape Characters, Markup Quotes */
-.code-highlighted:not(.dark) .hljs-built_in,
-.code-highlighted:not(.dark) .hljs-doctag, /* guessing */
-.code-highlighted:not(.dark) .hljs-quote,
-.code-highlighted:not(.dark) .hljs-keyword.hljs-atrule,
-.code-highlighted:not(.dark) .hljs-regexp {
-  color: #86c1b9;
-}
-
-/* base0D - Functions, Methods, Attribute IDs, Headings */
-.code-highlighted:not(.dark) .hljs-function .hljs-title,
-.code-highlighted:not(.dark) .hljs-attribute,
-.code-highlighted:not(.dark) .ruby .hljs-property,
-.code-highlighted:not(.dark) .hljs-title.function_,
-.code-highlighted:not(.dark) .hljs-section {
-  color: #7cafc2;
-}
-
-/* base0E - Keywords, Storage, Selector, Markup Italic, Diff Changed */
-.code-highlighted:not(.dark) .hljs-type,
-/* .hljs-selector-id, */
-/* .hljs-selector-class, */
-/* .hljs-selector-attr, */
-/* .hljs-selector-pseudo, */
-.code-highlighted:not(.dark) .hljs-template-tag,
-.code-highlighted:not(.dark) .diff .hljs-meta,
-.code-highlighted:not(.dark) .hljs-keyword {
-  color: #ba8baf;
-}
-.code-highlighted:not(.dark) .hljs-emphasis {
-  color: #ba8baf;
-  font-style: italic;
-}
-
-/* base0F - Deprecated, Opening/Closing Embedded Language Tags, e.g. <?php ?> */
-.code-highlighted:not(.dark) .hljs-meta,
-/*
-  prevent top level .keyword and .string scopes
-  from leaking into meta by accident
-*/
-.code-highlighted:not(.dark) .hljs-meta .hljs-keyword,
-.code-highlighted:not(.dark) .hljs-meta .hljs-string
-{
-  color: #a16946;
-}
-
-.code-highlighted:not(.dark) .hljs-meta .hljs-keyword,
-/* for v10 compatible themes */
-.code-highlighted:not(.dark) .hljs-meta-keyword {
-  font-weight: bold;
-}
