--- blocks/code/code.css
@@ -1,24 +1,0 @@
-.code code {
-  counter-reset: line;
-  padding: 0;
-}
-
-.code code span {
-  display: block;
-}
-
-.code code span.comment {
-  color: #999;
-}
-
-.code.line-numbers code span:before {
-  counter-increment: line;
-  content: counter(line);
-  display: inline-block;
-  padding: 0 .5em 0 0;
-  margin-right: .5em;
-  color: #ccc;
-  border-right: 1px solid #ccc;
-  min-width: 2em;
-  text-align: right;
-}
