+++ fstab.yaml
@@ -1,2 +1,2 @@
 mountpoints:
-  /: https://drive.google.com/drive/folders/1KhjvZubxYS5-mOxbRwLMN2OZtGH3nBVJ
+  /: https://drive.google.com/drive/u/0/folders/1Dmrg3_0_KUDoGb-HVraiaNv7W0D7TAHP
